Update index.html
new prompt given to chat gpt - Ok I have edits to the website I want to make. I want the header and footer to be the same color as the body. I also want the Body text to read "Alexander Horst and Sharvari Johari" on one line and under it "August 2024, New York City" and under that "more information coming soon". Each line should be centered and in smaller font than the one below it. That is the only text I want on the site. what should the html and css look like